Started new timer classes that will give a uniform interface across uno, mini, mega, and due.   Not complete, but getting close.  Most significant change to working code is the introduction TimverVectors that has a single declaration for each interupt vector.  I had too many classes that were declaring the same ISR macros.

+++ LampMatrix/scratchpins/timers/Alarm.cpp
@@ -30,7 +30,7 @@ Alarm& Alarm::detachInterrupt() {
 	return *this;
 }
 
-Alarm& Alarm::setPeriod(unsigned int periodMs) {
+Alarm& Alarm::setPeriod(unsigned long periodMs) {
 	// Convert period in microseconds to frequency in Hz
 	double frequency = 1000000.0 / periodMs;
 	setFrequency(frequency);
